# docs/swagger.yaml
basePath: /api/v1
definitions:
  atlas_internal_response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  atlas_internal_response.APIResponse-internal_modules_auth_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_auth.LoginResponse'
      message:
        type: string
      status:
        type: string
    type: object
  atlas_internal_response.APIResponse-internal_modules_auth_RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_auth.RefreshResponse'
      message:
        type: string
      status:
        type: string
    type: object
  atlas_internal_response.APIResponse-internal_modules_auth_RegistrationResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_auth.RegistrationResponse'
      message:
        type: string
      status:
        type: string
    type: object
  atlas_internal_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/atlas_internal_response.APIError'
      status:
        type: string
    type: object
  atlas_internal_response.MessageResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  internal_modules_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_modules_auth.LoginResponse:
    properties:
      tokens:
        $ref: '#/definitions/internal_modules_auth.Tokens'
      user:
        $ref: '#/definitions/internal_modules_auth.UserResponse'
    type: object
  internal_modules_auth.RefreshResponse:
    properties:
      tokens:
        $ref: '#/definitions/internal_modules_auth.Tokens'
    type: object
  internal_modules_auth.RegistrationRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    required:
    - confirm_password
    - email
    - first_name
    - last_name
    - password
    type: object
  internal_modules_auth.RegistrationResponse:
    properties:
      tokens:
        $ref: '#/definitions/internal_modules_auth.Tokens'
      user:
        $ref: '#/definitions/internal_modules_auth.UserResponse'
    type: object
  internal_modules_auth.Tokens:
    properties:
      access_token:
        type: string
    type: object
  internal_modules_auth.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      public_id:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for Atlas.
  title: Atlas API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Authenticates an email and password, opens a new session, and returns an access token.
        The refresh token is returned as an HttpOnly, SameSite=Strict cookie scoped to /api/v1/auth — it is never included in the response body.
        An unknown email and a wrong password return the same 401 so the endpoint cannot be used to discover registered addresses.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/atlas_internal_response.APIResponse-internal_modules_auth_LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      description: |-
        Revokes the current session and expires the refresh token cookie. Requires the refresh token cookie.
        Revoking an already-revoked session is a no-op success, so repeat calls are safe.
        Clients should clear local auth state and redirect regardless of the response status — logout is best-effort by design.
      produces:
      - application/json
      responses:
        "200":
          description: Session revoked and cookie cleared
          schema:
            $ref: '#/definitions/atlas_internal_response.MessageResponse'
        "401":
          description: Missing or invalid refresh token cookie
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
      summary: Log out
      tags:
      - auth
  /auth/refresh:
    post:
      description: |-
        Issues a new access token and rotates the refresh token cookie. Requires the refresh token cookie.
        The session id (sid) is preserved across refreshes, so access tokens issued earlier in the session remain valid until they expire on their own.

        NOTE — clients must serialize refresh calls. Every call rotates the refresh token, and presenting a token that has already been rotated away is treated as theft: the whole session is revoked and the user must log in again. Keep a single in-flight request to this endpoint and have all concurrent 401s await it. Firing two refreshes in parallel will log the user out.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/atlas_internal_response.APIResponse-internal_modules_auth_RefreshResponse'
        "401":
          description: Missing/invalid cookie, inactive session, or token reuse detected
            (session revoked)
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
      summary: Refresh the access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        Creates a user account, opens a session, and returns an access token.
        The refresh token is returned as an HttpOnly, SameSite=Strict cookie scoped to /api/v1/auth — it is never included in the response body.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.RegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/atlas_internal_response.APIResponse-internal_modules_auth_RegistrationResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
        "422":
          description: Passwords do not match, or password fails policy
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/atlas_internal_response.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"